requirement.txt
aniso8601==7.0.0
asgiref==3.4.1
Django==3.2.20
django-filter==21.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
graphene==2.1.9
graphene-django==2.16.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
psycopg2-binary==2.9.7
PyJWT==1.7.1
pytz==2023.3
Rx==1.6.3
serious-django-graphene==0.5.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.1.1
